renamed auto folder to auto routines
