+++ src/main/java/frc/robot/Robot.java
@@ -15,10 +15,10 @@ import edu.wpi.first.wpilibj2.command.Command;
 import edu.wpi.first.wpilibj2.command.CommandScheduler;
 import edu.wpi.first.wpilibj2.command.InstantCommand;
 import edu.wpi.first.wpilibj2.command.button.CommandXboxController;
-import frc.robot.auto.AutonomousBalanceMobile;
-import frc.robot.auto.AutonomousBalanceNoMobile;
-import frc.robot.auto.AutonomousDumpDoNothing;
-import frc.robot.auto.AutonomousGrab;
+import frc.robot.autoRoutines.AutonomousBalanceMobile;
+import frc.robot.autoRoutines.AutonomousBalanceNoMobile;
+import frc.robot.autoRoutines.AutonomousDumpDoNothing;
+import frc.robot.autoRoutines.AutonomousGrab;
 import frc.robot.commands.ArcadeDrive;
 import frc.robot.commands.IntakeToggle;
 import frc.robot.commands.LimelightCommand;
